requirements-dev.txt
# From our manifest.json for our custom component
asyncio>=3.4.3
sonnen_api_v2>=0.5.12
setuptools==75.1.0
voluptuous>=0.15.2

# Strictly for tests
pytest
pytest-asyncio
pytest-cov
pytest-mock
coverage>=7.2.1
pytest-homeassistant-custom-component
freezegun~=1.5.1
asyncmock>=0.4.2